ACOMMONS-14: Add SonarLintCache component
`SonarLintCache` is a component that can be used to transfer data in memory
between plugins in a SonarLint context, using the existing caching
mechanisms.

At the time of writing, this component used only by the DBD plugin, to consume
IRs produced by DBD custom rules.
We place it in analyzer-commons, since it is a depencency of multiple
language plugins (sonar-java and sonar-python).

The component is just an intermediate solution until a dedicated mechanism to
communicate between plugins with sufficient capabilities is available.